backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
openai==1.6.1
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0 